+++ src/engine/BilinearConstraint.h
@@ -1,20 +1,19 @@
 /*********************                                                        */
-/*! \file SigmoidConstraint.h
- ** \verbatim
- ** Top contributors (to current version):
- **   Teruhiro Tagomori, Haoze Wu
- ** This file is part of the Marabou project.
- ** Copyright (c) 2017-2019 by the authors listed in the file AUTHORS
- ** in the top-level source directory) and their institutional affiliations.
- ** All rights reserved. See the file COPYING in the top-level source
- ** directory for licensing information.\endverbatim
- **
- ** [[ Add lengthier description here ]]
-
+/*! \file BilinearConstraint.h
+** \verbatim
+** Top contributors (to current version):
+**   Andrew Wu
+** This file is part of the Marabou project.
+** Copyright (c) 2017-2019 by the authors listed in the file AUTHORS
+** in the top-level source directory) and their institutional affiliations.
+** All rights reserved. See the file COPYING in the top-level source
+** directory for licensing information.\endverbatim
+**
+** [[ Add lengthier description here ]]
 **/
 
-#ifndef __SigmoidConstraint_h__
-#define __SigmoidConstraint_h__
+#ifndef __BilinearConstraint_h__
+#define __BilinearConstraint_h__
 
 #include "List.h"
 #include "Map.h"
@@ -22,11 +21,11 @@
 
 #include <cmath>
 
-class SigmoidConstraint : public NonlinearConstraint
+class BilinearConstraint : public NonlinearConstraint
 {
 public:
-    SigmoidConstraint( unsigned b, unsigned f );
-    SigmoidConstraint( const String &serializedSigmoid );
+    BilinearConstraint( unsigned b1, unsigned b2, unsigned f );
+    BilinearConstraint( const String &serializedBilinear );
 
     /*
       Get the type of this constraint.
@@ -41,11 +40,11 @@ public:
     /*
       Restore the state of this constraint from the given one.
     */
-    void restoreState( const NonlinearConstraint *state ) override; 
+    void restoreState( const NonlinearConstraint *state ) override;
 
     /*
       Register/unregister the constraint with a talbeau.
-     */
+    */
     void registerAsWatcher( ITableau *tableau ) override;
     void unregisterAsWatcher( ITableau *tableau ) override;
 
@@ -69,15 +68,13 @@ public:
     /*
       Preprocessing related functions, to inform that a variable has
       been eliminated completely because it was fixed to some value,
-      or that a variable's index has changed (e.g., x4 is now called
+      or thisat a variable's index has changed (e.g., x4 is now called
       x2). constraintObsolete() returns true iff and the constraint
       has become obsolote as a result of variable eliminations.
     */
     void eliminateVariable( unsigned variable, double fixedValue ) override;
     void updateVariableIndex( unsigned oldIndex, unsigned newIndex ) override;
     bool constraintObsolete() const override;
-
-    bool supportVariableElimination() const override { return true; };
 
     /*
       Get the tightenings entailed by the constraint.
@@ -90,11 +87,6 @@ public:
     bool satisfied() const override;
 
     /*
-      Dump the current state of the constraint.
-    */
-    void dump( String &output ) const override;
-
-    /*
       Returns string with shape: sigmoid, _f, _b
     */
     String serializeToString() const override;
@@ -102,27 +94,11 @@ public:
     /*
       Get the index of the B and F variables.
     */
-    unsigned getB() const;
+    Vector<unsigned> getBs() const;
     unsigned getF() const;
 
-    /*
-      Compute the sigmoid function.
-    */
-    static double sigmoid( double x );
-
-    /*
-      Compute the inverse of the sigmoid function.
-    */
-    static double sigmoidInverse( double y );
-
-    /*
-      Compute the derivative of the sigmoid function.
-    */
-    static double sigmoidDerivative( double x );
-
 private:
-    unsigned _b, _f;
-    bool _haveEliminatedVariables;
+    unsigned _b1, _b2, _f;
 };
 
-#endif // __SigmoidConstraint_h__
+#endif // __BilinearConstraint_h__