+++ src/input_parsers/NetworkParser.h
@@ -30,7 +30,7 @@
 #include "PiecewiseLinearConstraint.h"
 #include "SigmoidConstraint.h"
 #include "SignConstraint.h"
-#include "TranscendentalConstraint.h"
+#include "NonlinearConstraint.h"
 #include <utility>
 
 typedef unsigned int Variable;