+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Added support for onnx networks with tanh nodes to command-line interface.
 * Added proof producing versions of Sign, Max, Absolute Value and Disjunction constraints.
 * Added support for properties provided in VNN-LIB format for ONNX networks via the Python API.
-* Support additional non-linear constraints Softmax and Bilinear
+* Support additional non-linear constraints Softmax and Bilinear.
 * Added support for Pythonic API.
 
 ## Version 1.0.0